+++ include/erl_geometry/octree_impl.hpp
@@ -111,7 +111,7 @@ namespace erl::geometry {
             return m_tree_size_;
         }
 
-        [[maybe_unused]] [[nodiscard]] Eigen::Vector3d
+        [[nodiscard]] Eigen::Vector3d
         GetTreeCenter() const {
             double x = this->KeyToCoord(m_tree_key_offset_);
             return {x, x, x};
@@ -123,7 +123,7 @@ namespace erl::geometry {
             return {key, key, key};
         }
 
-        [[maybe_unused]] [[nodiscard]] Eigen::Vector3d
+        [[nodiscard]] Eigen::Vector3d
         GetTreeMaxHalfSize() const {
             double size = -this->KeyToCoord(0);
             return {size, size, size};
@@ -184,7 +184,7 @@ namespace erl::geometry {
             }
 
             min_x = min_y = min_z = std::numeric_limits<double>::infinity();
-            for (auto it = this->begin(), end = this->end(); it != end; ++it) {
+            for (auto it = this->BeginLeaf(), end = this->EndLeaf(); it != end; ++it) {
                 const double half_size = it.GetNodeSize() / 2.;
                 const double node_min_x = it.GetX() - half_size;
                 const double node_min_y = it.GetY() - half_size;
@@ -220,7 +220,7 @@ namespace erl::geometry {
             }
 
             max_x = max_y = max_z = -std::numeric_limits<double>::infinity();
-            for (auto it = this->begin(), end = this->end(); it != end; ++it) {
+            for (auto it = this->BeginLeaf(), end = this->EndLeaf(); it != end; ++it) {
                 double half_size = it.GetNodeSize() / 2.;
                 const double node_max_x = it.GetX() + half_size;
                 const double node_max_y = it.GetY() + half_size;
@@ -263,7 +263,7 @@ namespace erl::geometry {
 
             min_x = min_y = min_z = std::numeric_limits<double>::infinity();
             max_x = max_y = max_z = -std::numeric_limits<double>::infinity();
-            for (auto it = this->begin(), end = this->end(); it != end; ++it) {
+            for (auto it = this->BeginLeaf(), end = this->EndLeaf(); it != end; ++it) {
                 const double size = it.GetNodeSize();
                 const double half_size = size / 2.;
                 const double node_max_x = it.GetX() + half_size;
@@ -308,7 +308,7 @@ namespace erl::geometry {
 
             double min_x = std::numeric_limits<double>::infinity(), min_y = min_x, min_z = min_x;
             double max_x = -std::numeric_limits<double>::infinity(), max_y = max_x, max_z = max_x;
-            for (auto it = this->begin(), end = this->end(); it != end; ++it) {
+            for (auto it = this->BeginLeaf(), end = this->EndLeaf(); it != end; ++it) {
                 const double size = it.GetNodeSize();
                 const double half_size = size / 2.;
                 const double node_max_x = it.GetX() + half_size;
@@ -366,7 +366,7 @@ namespace erl::geometry {
             return num_leaf_nodes;
         }
 
-        [[maybe_unused]] [[nodiscard]] std::size_t
+        [[nodiscard]] std::size_t
         GetMemoryUsage() const override {
             const std::size_t number_of_leaf_nodes = this->ComputeNumberOfLeafNodes();
             const std::size_t number_of_inner_nodes = m_tree_size_ - number_of_leaf_nodes;
@@ -734,7 +734,7 @@ namespace erl::geometry {
         }
 
         //-- iterator implementation
-        class IteratorBase {
+        class IteratorBase : public AbstractOctree::OctreeNodeIterator {
         public:
             struct StackElement {
                 Node *node = nullptr;
@@ -820,39 +820,54 @@ namespace erl::geometry {
             }
 
             [[nodiscard]] double
-            GetX() const {
+            GetX() const override {
                 const StackElement &top = m_stack_.back();
                 return m_tree_->KeyToCoord(top.key[0], top.node->GetDepth());
             }
 
             [[nodiscard]] double
-            GetY() const {
+            GetY() const override {
                 const StackElement &top = m_stack_.back();
                 return m_tree_->KeyToCoord(top.key[1], top.node->GetDepth());
             }
 
             [[nodiscard]] double
-            GetZ() const {
+            GetZ() const override {
                 const StackElement &top = m_stack_.back();
                 return m_tree_->KeyToCoord(top.key[2], top.node->GetDepth());
             }
 
             [[nodiscard]] double
-            GetNodeSize() const {
+            GetNodeSize() const override {
                 return m_tree_->GetNodeSize(m_stack_.back().node->GetDepth());
             }
 
-            [[maybe_unused]] [[nodiscard]] Aabb3D
+            [[nodiscard]] uint32_t
+            GetDepth() const override {
+                return m_stack_.back().node->GetDepth();
+            }
+
+            [[nodiscard]] const AbstractOctreeNode *
+            GetNode() override {
+                return m_stack_.back().node;
+            }
+
+            [[nodiscard]] bool
+            IsValid() const override {
+                return !m_stack_.empty();
+            }
+
+            [[nodiscard]] Aabb3D
             GetNodeAabb() const {
                 return m_tree_->GetNodeAabb(m_stack_.back().key, m_stack_.back().node->GetDepth());
             }
 
-            [[maybe_unused]] [[nodiscard]] const OctreeKey &
+            [[nodiscard]] const OctreeKey &
             GetKey() const {
                 return m_stack_.back().key;
             }
 
-            [[maybe_unused]] [[nodiscard]] OctreeKey
+            [[nodiscard]] OctreeKey
             GetIndexKey() const {
                 const StackElement &top = m_stack_.back();
                 return m_tree_->AdjustKeyToDepth(top.key, top.node->GetDepth());
@@ -949,6 +964,11 @@ namespace erl::geometry {
                 if (!this->m_stack_.empty()) { this->SingleIncrement1(); }
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
+            }
+
+            void
+            Next() override {
+                ++(*this);
             }
         };
 
@@ -1063,6 +1083,11 @@ namespace erl::geometry {
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
             }
+
+            void
+            Next() override {
+                ++(*this);
+            }
         };
 
         /**
@@ -1112,6 +1137,11 @@ namespace erl::geometry {
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
             }
+
+            void
+            Next() override {
+                ++(*this);
+            }
         };
 
         class LeafIterator : public IteratorBase {
@@ -1145,6 +1175,11 @@ namespace erl::geometry {
                 do { this->SingleIncrement1(); } while (!this->m_stack_.empty() && !this->IsLeaf());
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
+            }
+
+            void
+            Next() override {
+                ++(*this);
             }
         };
 
@@ -1193,6 +1228,11 @@ namespace erl::geometry {
                 do { this->SingleIncrement1(); } while (!this->m_stack_.empty() && !this->IsLeaf());
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
+            }
+
+            void
+            Next() override {
+                ++(*this);
             }
         };
 
@@ -1339,6 +1379,11 @@ namespace erl::geometry {
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
             }
+
+            void
+            Next() override {
+                ++(*this);
+            }
         };
 
         class EastLeafNeighborIterator : public LeafNeighborIteratorBase {
@@ -1386,6 +1431,11 @@ namespace erl::geometry {
                 if (!this->m_stack_.empty()) { this->SingleIncrementOf(sk_ChangingDim1_, sk_ChangingDim2_); }
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
+            }
+
+            void
+            Next() override {
+                ++(*this);
             }
         };
 
@@ -1436,6 +1486,11 @@ namespace erl::geometry {
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
             }
+
+            void
+            Next() override {
+                ++(*this);
+            }
         };
 
         class SouthLeafNeighborIterator : public LeafNeighborIteratorBase {
@@ -1483,6 +1538,11 @@ namespace erl::geometry {
                 if (!this->m_stack_.empty()) { this->SingleIncrementOf(sk_ChangingDim1_, sk_ChangingDim2_); }
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
+            }
+
+            void
+            Next() override {
+                ++(*this);
             }
         };
 
@@ -1533,6 +1593,11 @@ namespace erl::geometry {
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
             }
+
+            void
+            Next() override {
+                ++(*this);
+            }
         };
 
         class BottomLeafNeighborIterator : public LeafNeighborIteratorBase {
@@ -1581,6 +1646,11 @@ namespace erl::geometry {
                 if (!this->m_stack_.empty()) { this->SingleIncrementOf(sk_ChangingDim1_, sk_ChangingDim2_); }
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
+            }
+
+            void
+            Next() override {
+                ++(*this);
             }
         };
 
@@ -1638,7 +1708,7 @@ namespace erl::geometry {
                 if (this->m_stack_.empty()) { this->Terminate(); }
             }
 
-            [[maybe_unused]] [[nodiscard]] double
+            [[nodiscard]] double
             GetDistance() const {
                 if (this->m_stack_.empty()) { return 0.; }
                 return *reinterpret_cast<double *>(this->m_stack_.back().data.get());
@@ -1658,6 +1728,11 @@ namespace erl::geometry {
                 if (!this->m_stack_.empty()) { this->SingleIncrement2(); }
                 if (this->m_stack_.empty()) { this->Terminate(); }
                 return *this;
+            }
+
+            void
+            Next() override {
+                ++(*this);
             }
 
         protected:
@@ -1756,24 +1831,6 @@ namespace erl::geometry {
          * @return
          */
         [[nodiscard]] LeafIterator
-        begin(uint32_t max_depth = 0) const {
-            return LeafIterator(this, max_depth);
-        }
-
-        /**
-         * End iterator of LeafIterator.
-         */
-        [[nodiscard]] LeafIterator
-        end() const {
-            return LeafIterator();
-        }
-
-        /**
-         * Iterator that iterates over all leaf nodes.
-         * @param max_depth
-         * @return
-         */
-        [[nodiscard]] LeafIterator
         BeginLeaf(uint32_t max_depth = 0) const {
             return LeafIterator(this, max_depth);
         }
@@ -1786,12 +1843,17 @@ namespace erl::geometry {
             return LeafIterator();
         }
 
-        [[maybe_unused]] [[nodiscard]] LeafOfNodeIterator
+        [[nodiscard]] std::shared_ptr<AbstractOctree::OctreeNodeIterator>
+        GetLeafIterator(uint32_t max_depth) const override {
+            return std::make_shared<LeafIterator>(this, max_depth);
+        }
+
+        [[nodiscard]] LeafOfNodeIterator
         BeginLeafOfNode(const OctreeKey &key, uint32_t node_depth, uint32_t max_depth = 0) const {
             return LeafOfNodeIterator(key, node_depth, this, max_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] LeafOfNodeIterator
+        [[nodiscard]] LeafOfNodeIterator
         EndLeafOfNode() const {
             return LeafOfNodeIterator();
         }
@@ -1823,6 +1885,19 @@ namespace erl::geometry {
             return LeafInAabbIterator();
         }
 
+        [[nodiscard]] std::shared_ptr<AbstractOctree::OctreeNodeIterator>
+        GetLeafInAabbIterator(const Aabb3D &aabb, uint32_t max_depth) const {
+            return std::make_shared<LeafInAabbIterator>(
+                aabb.min().x(),
+                aabb.min().y(),
+                aabb.min().z(),
+                aabb.max().x(),
+                aabb.max().y(),
+                aabb.max().z(),
+                this,
+                max_depth);
+        }
+
         [[nodiscard]] TreeIterator
         BeginTree(uint32_t max_depth = 0) const {
             return TreeIterator(this, max_depth);
@@ -1831,6 +1906,11 @@ namespace erl::geometry {
         [[nodiscard]] TreeIterator
         EndTree() const {
             return TreeIterator();
+        }
+
+        [[nodiscard]] std::shared_ptr<AbstractOctree::OctreeNodeIterator>
+        GetTreeIterator(const uint32_t max_depth) const override {
+            return std::make_shared<TreeIterator>(this, max_depth);
         }
 
         [[nodiscard]] TreeInAabbIterator
@@ -1860,97 +1940,110 @@ namespace erl::geometry {
             return TreeInAabbIterator();
         }
 
-        [[maybe_unused]] [[nodiscard]] WestLeafNeighborIterator
+        [[nodiscard]] std::shared_ptr<AbstractOctree::OctreeNodeIterator>
+        GetTreeInAabbIterator(const Aabb3D &aabb, uint32_t max_depth) const override {
+            return std::make_shared<TreeInAabbIterator>(
+                aabb.min().x(),
+                aabb.min().y(),
+                aabb.min().z(),
+                aabb.max().x(),
+                aabb.max().y(),
+                aabb.max().z(),
+                this,
+                max_depth);
+        }
+
+        [[nodiscard]] WestLeafNeighborIterator
         BeginWestLeafNeighbor(double x, double y, double z, uint32_t max_leaf_depth = 0) const {
             return WestLeafNeighborIterator(x, y, z, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] WestLeafNeighborIterator
+        [[nodiscard]] WestLeafNeighborIterator
         BeginWestLeafNeighbor(const OctreeKey &key, uint32_t key_depth, uint32_t max_leaf_depth = 0) const {
             return WestLeafNeighborIterator(key, key_depth, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] WestLeafNeighborIterator
+        [[nodiscard]] WestLeafNeighborIterator
         EndWestLeafNeighbor() const {
             return WestLeafNeighborIterator();
         }
 
-        [[maybe_unused]] [[nodiscard]] EastLeafNeighborIterator
+        [[nodiscard]] EastLeafNeighborIterator
         BeginEastLeafNeighbor(double x, double y, double z, uint32_t max_leaf_depth = 0) const {
             return EastLeafNeighborIterator(x, y, z, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] EastLeafNeighborIterator
+        [[nodiscard]] EastLeafNeighborIterator
         BeginEastLeafNeighbor(const OctreeKey &key, uint32_t key_depth, uint32_t max_leaf_depth = 0) const {
             return EastLeafNeighborIterator(key, key_depth, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] EastLeafNeighborIterator
+        [[nodiscard]] EastLeafNeighborIterator
         EndEastLeafNeighbor() const {
             return EastLeafNeighborIterator();
         }
 
-        [[maybe_unused]] [[nodiscard]] NorthLeafNeighborIterator
+        [[nodiscard]] NorthLeafNeighborIterator
         BeginNorthLeafNeighbor(double x, double y, double z, uint32_t max_leaf_depth = 0) const {
             return NorthLeafNeighborIterator(x, y, z, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] NorthLeafNeighborIterator
+        [[nodiscard]] NorthLeafNeighborIterator
         BeginNorthLeafNeighbor(const OctreeKey &key, uint32_t key_depth, uint32_t max_leaf_depth = 0) const {
             return NorthLeafNeighborIterator(key, key_depth, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] NorthLeafNeighborIterator
+        [[nodiscard]] NorthLeafNeighborIterator
         EndNorthLeafNeighbor() const {
             return NorthLeafNeighborIterator();
         }
 
-        [[maybe_unused]] [[nodiscard]] SouthLeafNeighborIterator
+        [[nodiscard]] SouthLeafNeighborIterator
         BeginSouthLeafNeighbor(double x, double y, double z, uint32_t max_leaf_depth = 0) const {
             return SouthLeafNeighborIterator(x, y, z, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] SouthLeafNeighborIterator
+        [[nodiscard]] SouthLeafNeighborIterator
         BeginSouthLeafNeighbor(const OctreeKey &key, uint32_t key_depth, uint32_t max_leaf_depth = 0) const {
             return SouthLeafNeighborIterator(key, key_depth, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] SouthLeafNeighborIterator
+        [[nodiscard]] SouthLeafNeighborIterator
         EndSouthLeafNeighbor() const {
             return SouthLeafNeighborIterator();
         }
 
-        [[maybe_unused]] [[nodiscard]] TopLeafNeighborIterator
+        [[nodiscard]] TopLeafNeighborIterator
         BeginTopLeafNeighbor(double x, double y, double z, uint32_t max_leaf_depth = 0) const {
             return TopLeafNeighborIterator(x, y, z, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] TopLeafNeighborIterator
+        [[nodiscard]] TopLeafNeighborIterator
         BeginTopLeafNeighbor(const OctreeKey &key, uint32_t key_depth, uint32_t max_leaf_depth = 0) const {
             return TopLeafNeighborIterator(key, key_depth, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] TopLeafNeighborIterator
+        [[nodiscard]] TopLeafNeighborIterator
         EndTopLeafNeighbor() const {
             return TopLeafNeighborIterator();
         }
 
-        [[maybe_unused]] [[nodiscard]] BottomLeafNeighborIterator
+        [[nodiscard]] BottomLeafNeighborIterator
         BeginBottomLeafNeighbor(double x, double y, double z, uint32_t max_leaf_depth = 0) const {
             return BottomLeafNeighborIterator(x, y, z, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] BottomLeafNeighborIterator
+        [[nodiscard]] BottomLeafNeighborIterator
         BeginBottomLeafNeighbor(const OctreeKey &key, uint32_t key_depth, uint32_t max_leaf_depth = 0) const {
             return BottomLeafNeighborIterator(key, key_depth, this, max_leaf_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] BottomLeafNeighborIterator
+        [[nodiscard]] BottomLeafNeighborIterator
         EndBottomLeafNeighbor() const {
             return BottomLeafNeighborIterator();
         }
 
-        [[maybe_unused]] [[nodiscard]] NodeOnRayIterator
+        [[nodiscard]] NodeOnRayIterator
         BeginNodeOnRay(
             double px,
             double py,
@@ -1966,9 +2059,25 @@ namespace erl::geometry {
             return NodeOnRayIterator(px, py, pz, vx, vy, vz, max_range, bidirectional, /*tree*/ this, leaf_only, min_node_depth, max_node_depth);
         }
 
-        [[maybe_unused]] [[nodiscard]] NodeOnRayIterator
+        [[nodiscard]] NodeOnRayIterator
         EndNodeOnRay() const {
             return {};
+        }
+
+        [[nodiscard]] std::shared_ptr<AbstractOctree::OctreeNodeIterator>
+        GetNodeOnRayIterator(
+            double px,
+            double py,
+            double pz,
+            double vx,
+            double vy,
+            double vz,
+            double max_range,
+            bool bidirectional,
+            bool leaf_only,
+            uint32_t min_node_depth,
+            uint32_t max_node_depth) const override {
+            return std::make_shared<NodeOnRayIterator>(px, py, pz, vx, vy, vz, max_range, bidirectional, this, leaf_only, min_node_depth, max_node_depth);
         }
 
         //-- ray tracing
@@ -2287,7 +2396,11 @@ namespace erl::geometry {
          * @param key key of the child node
          */
         virtual void
-        OnDeleteNodeChild([[maybe_unused]] Node *node, [[maybe_unused]] Node *child, [[maybe_unused]] const OctreeKey &key) {}
+        OnDeleteNodeChild(Node *node, Node *child, const OctreeKey &key) {
+            (void) node;
+            (void) child;
+            (void) key;
+        }
 
         /**
          * Delete child nodes down to max_depth matching the given key of the given node that is at the given depth.
@@ -2473,7 +2586,7 @@ namespace erl::geometry {
             return node;
         }
 
-        [[maybe_unused]] Node *
+        Node *
         InsertNode(double x, double y, double z, const uint32_t depth = 0) {
             OctreeKey key;
             if (!CoordToKeyChecked(x, y, z, key)) {
@@ -2599,7 +2712,7 @@ namespace erl::geometry {
             m_metric_min_[0] = m_metric_min_[1] = m_metric_min_[2] = std::numeric_limits<double>::infinity();
             m_metric_max_[0] = m_metric_max_[1] = m_metric_max_[2] = -std::numeric_limits<double>::infinity();
 
-            for (auto it = this->begin(), end = this->end(); it != end; ++it) {
+            for (auto it = this->BeginLeaf(), end = this->EndLeaf(); it != end; ++it) {
                 const double size = it.GetNodeSize();
                 const double half_size = size / 2.0;
                 double x = it.GetX() - half_size;